+++ tests/test_audible_m4b_files.py
@@ -4,8 +4,6 @@ plugin, when the beets-audible plugin is loaded.
 
 import logging
 
-from typing import Optional
-
 from beets import config
 
 from tests.helper import FiletoteTestCase, MediaSetup
@@ -18,7 +16,7 @@ class FiletoteM4BFilesIgnoredTest(FiletoteTestCase):
     beets-audible plugin is present.
     """
 
-    def setUp(self, _other_plugins: Optional[list[str]] = None) -> None:
+    def setUp(self, _other_plugins: list[str] | None = None) -> None:
         """Provides shared setup for tests."""
         super().setUp(other_plugins=["audible"])