+++ tests/test_exclude.py
@@ -4,8 +4,6 @@ plugin.
 
 import os
 
-from typing import Optional
-
 import beets
 
 from beets import config
@@ -18,7 +16,7 @@ class FiletoteExcludeTest(FiletoteTestCase):
     beets-filetote plugin.
     """
 
-    def setUp(self, _other_plugins: Optional[list[str]] = None) -> None:
+    def setUp(self, _other_plugins: list[str] | None = None) -> None:
         """Provides shared setup for tests."""
         super().setUp()