+++ tests/test_printignored.py
@@ -1,7 +1,5 @@
 """Tests print ignored for the beets-filetote plugin."""
 
-from typing import Optional
-
 from beets import config
 
 from tests.helper import FiletoteTestCase, capture_log_with_traceback
@@ -10,7 +8,7 @@ from tests.helper import FiletoteTestCase, capture_log_with_traceback
 class FiletotePrintIgnoredTest(FiletoteTestCase):
     """Tests to check print ignored files functionality and configuration."""
 
-    def setUp(self, _other_plugins: Optional[list[str]] = None) -> None:
+    def setUp(self, _other_plugins: list[str] | None = None) -> None:
         """Provides shared setup for tests."""
         super().setUp()