+++ tests/test_rename_item_fields.py
@@ -2,8 +2,6 @@
 
 import logging
 
-from typing import Optional
-
 from beets import config
 
 from tests.helper import FiletoteTestCase
@@ -16,7 +14,7 @@ class FiletoteRenameItemFieldsTest(FiletoteTestCase):
     expected for custom path formats.
     """
 
-    def setUp(self, _other_plugins: Optional[list[str]] = None) -> None:
+    def setUp(self, _other_plugins: list[str] | None = None) -> None:
         """Provides shared setup for tests."""
         super().setUp()